rozetka_mini/pyproject.toml
[tool.poetry]
name = "rozetka-mini"
version = "0.2.0"
description = "Mini marketplace like Rozetka with Django + HTMX (luxe theme)"
authors = ["You <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
django-filter = "^24.3"
django-render-block = "^0.9"
django-htmx = "^1.17.3"
pillow = "^10.4.0"
django-environ = "^0.11.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.9"
isort = "^5.13.2"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
factory_boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
select = ["E","F","I","B"]
